+++ src/quacc/schemas/_aliases/vasp.py
@@ -56,12 +56,20 @@ class DoubleRelaxSchema(VaspSchema):
     """Type hint associated with double relaxation jobs."""
 
     relax1: VaspSchema
+    relax2: VaspSchema
 
 
 class MPMetaGGARelaxFlowSchema(VaspSchema):
     """Type hint associated with the MP meta-GGA relaxation flows."""
 
     prerelax: VaspSchema
+    relax: VaspSchema
+    static: VaspSchema
+
+
+class MPGGARelaxFlowSchema(VaspSchema):
+    """Type hint associated with the MP GGA relaxation flows."""
+
     relax: VaspSchema
     static: VaspSchema