+++ tests/core/recipes/vasp_recipes/mocked/test_vasp_recipes.py
@@ -4,6 +4,9 @@ from ase.build import bulk, molecule
 from quacc import SETTINGS
 from quacc.recipes.vasp.core import double_relax_job, relax_job, static_job
 from quacc.recipes.vasp.mp import (
+    mp_gga_relax_flow,
+    mp_gga_relax_job,
+    mp_gga_static_job,
     mp_metagga_prerelax_job,
     mp_metagga_relax_flow,
     mp_metagga_relax_job,
@@ -106,13 +109,13 @@ def test_doublerelax_job(tmp_path, monkeypatch):
     assert output["relax1"]["parameters"]["nsw"] > 0
     assert output["relax1"]["parameters"]["isif"] == 3
     assert output["relax1"]["parameters"]["lwave"] is False
-    assert output["nsites"] == len(atoms)
-    assert output["parameters"]["isym"] == 0
-    assert output["parameters"]["nsw"] > 0
-    assert output["parameters"]["isif"] == 3
-    assert output["parameters"]["lwave"] is False
     assert output["relax1"]["parameters"]["encut"] == 520
-    assert output["parameters"]["encut"] == 520
+    assert output["relax2"]["nsites"] == len(atoms)
+    assert output["relax2"]["parameters"]["isym"] == 0
+    assert output["relax2"]["parameters"]["nsw"] > 0
+    assert output["relax2"]["parameters"]["isif"] == 3
+    assert output["relax2"]["parameters"]["lwave"] is False
+    assert output["relax2"]["parameters"]["encut"] == 520
 
     output = double_relax_job(atoms, relax2_kwargs={"nelmin": 6})
     assert output["relax1"]["nsites"] == len(atoms)
@@ -120,30 +123,30 @@ def test_doublerelax_job(tmp_path, monkeypatch):
     assert output["relax1"]["parameters"]["nsw"] > 0
     assert output["relax1"]["parameters"]["isif"] == 3
     assert output["relax1"]["parameters"]["lwave"] is False
-    assert output["nsites"] == len(atoms)
-    assert output["parameters"]["isym"] == 0
-    assert output["parameters"]["nsw"] > 0
-    assert output["parameters"]["isif"] == 3
-    assert output["parameters"]["lwave"] is False
     assert output["relax1"]["parameters"]["encut"] == 520
-    assert output["parameters"]["encut"] == 520
-    assert output["parameters"]["nelmin"] == 6
+    assert output["relax2"]["nsites"] == len(atoms)
+    assert output["relax2"]["parameters"]["isym"] == 0
+    assert output["relax2"]["parameters"]["nsw"] > 0
+    assert output["relax2"]["parameters"]["isif"] == 3
+    assert output["relax2"]["parameters"]["lwave"] is False
+    assert output["relax2"]["parameters"]["encut"] == 520
+    assert output["relax2"]["parameters"]["nelmin"] == 6
 
     output = double_relax_job(atoms, relax_cell=False)
     assert output["relax1"]["nsites"] == len(atoms)
     assert output["relax1"]["parameters"]["isym"] == 0
     assert output["relax1"]["parameters"]["nsw"] > 0
     assert output["relax1"]["parameters"]["lwave"] is False
-    assert output["nsites"] == len(atoms)
-    assert output["parameters"]["isym"] == 0
-    assert output["parameters"]["nsw"] > 0
-    assert output["parameters"]["lwave"] is False
-    assert output["relax1"]["parameters"]["encut"] == 520
-    assert output["parameters"]["encut"] == 520
     assert output["relax1"]["parameters"]["isif"] == 2
-    assert output["parameters"]["isif"] == 2
+    assert output["relax1"]["parameters"]["encut"] == 520
+    assert output["relax2"]["nsites"] == len(atoms)
+    assert output["relax2"]["parameters"]["isym"] == 0
+    assert output["relax2"]["parameters"]["nsw"] > 0
+    assert output["relax2"]["parameters"]["lwave"] is False
+    assert output["relax2"]["parameters"]["encut"] == 520
+    assert output["relax2"]["parameters"]["isif"] == 2
 
-    double_relax_job(atoms, relax1_kwargs={"kpts": [1, 1, 1]})
+    assert double_relax_job(atoms, relax1_kwargs={"kpts": [1, 1, 1]})
 
 
 def test_slab_static_job(tmp_path, monkeypatch):
@@ -363,13 +366,13 @@ def test_mp_metagga_prerelax_job(tmp_path, monkeypatch):
     output = mp_metagga_prerelax_job(atoms)
     assert output["nsites"] == len(atoms)
     assert output["parameters"] == {
-        "algo": "All",
+        "algo": "all",
         "ediff": 1e-5,
         "ediffg": -0.05,
         "efermi": "midgap",  # added by copilot
         "enaug": 1360,
         "encut": 680,
-        "gga": "PS",
+        "gga": "ps",
         "ibrion": 2,
         "isif": 3,
         "ismear": 0,
@@ -381,38 +384,38 @@ def test_mp_metagga_prerelax_job(tmp_path, monkeypatch):
         "lelf": False,
         "lmixtau": True,
         "lorbit": 11,
-        "lreal": "Auto",
+        "lreal": "auto",
         "lvtot": False,  # disabled by us
         "lwave": True,
         "magmom": [0.6],
         "nelm": 200,
         "nsw": 99,
-        "prec": "Accurate",
+        "prec": "accurate",
         "setups": {"Al": ""},
         "sigma": 0.05,
-        "pp": "PBE",
+        "pp": "pbe",
     }
 
     output = mp_metagga_prerelax_job(atoms, bandgap=0)
     assert output["nsites"] == len(atoms)
-    assert output["parameters"]["gga"] == "PS"
+    assert output["parameters"]["gga"] == "ps"
     assert output["parameters"]["ediffg"] == -0.05
     assert output["parameters"]["encut"] == 680
     assert output["parameters"]["kspacing"] == 0.22
     assert output["parameters"]["ismear"] == 0
     assert output["parameters"]["sigma"] == 0.05
-    assert output["parameters"]["pp"] == "PBE"
+    assert output["parameters"]["pp"] == "pbe"
     assert "metagga" not in output["parameters"]
 
     output = mp_metagga_prerelax_job(atoms, bandgap=100)
     assert output["nsites"] == len(atoms)
-    assert output["parameters"]["gga"] == "PS"
+    assert output["parameters"]["gga"] == "ps"
     assert output["parameters"]["ediffg"] == -0.05
     assert output["parameters"]["encut"] == 680
     assert output["parameters"]["kspacing"] == 0.44
     assert output["parameters"]["ismear"] == 0
     assert output["parameters"]["sigma"] == 0.05
-    assert output["parameters"]["pp"] == "PBE"
+    assert output["parameters"]["pp"] == "pbe"
     assert "metagga" not in output["parameters"]
 
 
@@ -422,9 +425,9 @@ def test_mp_metagga_relax_job(tmp_path, monkeypatch):
     atoms = bulk("Al")
 
     output = mp_metagga_relax_job(atoms)
-    assert output["nsites"] == len(atoms)
-    assert output["parameters"] == {
-        "algo": "All",
+    assert output["relax2"]["nsites"] == len(atoms)
+    assert output["relax2"]["parameters"] == {
+        "algo": "all",
         "ediff": 1e-5,
         "ediffg": -0.02,
         "efermi": "midgap",  # added by copilot
@@ -441,38 +444,38 @@ def test_mp_metagga_relax_job(tmp_path, monkeypatch):
         "lelf": False,
         "lmixtau": True,
         "lorbit": 11,
-        "lreal": "Auto",
+        "lreal": "auto",
         "lvtot": False,  # disabled by us
         "lwave": True,
         "magmom": [0.6],
-        "metagga": "R2scan",
+        "metagga": "r2scan",
         "nelm": 200,
         "nsw": 99,
-        "prec": "Accurate",
+        "prec": "accurate",
         "sigma": 0.05,
-        "pp": "PBE",
+        "pp": "pbe",
         "setups": {"Al": ""},
     }
 
     output = mp_metagga_relax_job(atoms, bandgap=0)
-    assert output["nsites"] == len(atoms)
-    assert output["parameters"]["metagga"].lower() == "r2scan"
-    assert output["parameters"]["ediffg"] == -0.02
-    assert output["parameters"]["encut"] == 680
-    assert output["parameters"]["kspacing"] == 0.22
-    assert output["parameters"]["ismear"] == 0
-    assert output["parameters"]["sigma"] == 0.05
-    assert output["parameters"]["pp"] == "PBE"
+    assert output["relax2"]["nsites"] == len(atoms)
+    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax2"]["parameters"]["encut"] == 680
+    assert output["relax2"]["parameters"]["kspacing"] == 0.22
+    assert output["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax2"]["parameters"]["sigma"] == 0.05
+    assert output["relax2"]["parameters"]["pp"] == "pbe"
 
     output = mp_metagga_relax_job(atoms, bandgap=100)
-    assert output["nsites"] == len(atoms)
-    assert output["parameters"]["metagga"].lower() == "r2scan"
-    assert output["parameters"]["ediffg"] == -0.02
-    assert output["parameters"]["encut"] == 680
-    assert output["parameters"]["kspacing"] == 0.44
-    assert output["parameters"]["ismear"] == 0
-    assert output["parameters"]["sigma"] == 0.05
-    assert output["parameters"]["pp"] == "PBE"
+    assert output["relax2"]["nsites"] == len(atoms)
+    assert output["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax2"]["parameters"]["encut"] == 680
+    assert output["relax2"]["parameters"]["kspacing"] == 0.44
+    assert output["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax2"]["parameters"]["sigma"] == 0.05
+    assert output["relax2"]["parameters"]["pp"] == "pbe"
 
 
 def test_mp_metagga_static_job(tmp_path, monkeypatch):
@@ -501,12 +504,12 @@ def test_mp_metagga_static_job(tmp_path, monkeypatch):
         "lvtot": True,
         "lwave": True,  # enabled by us
         "magmom": [0.6],
-        "metagga": "R2scan",
+        "metagga": "r2scan",
         "nelm": 200,
         "nsw": 0,
-        "prec": "Accurate",
+        "prec": "accurate",
         "sigma": 0.05,
-        "pp": "PBE",
+        "pp": "pbe",
         "setups": {"Al": ""},
     }
 
@@ -518,30 +521,30 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
 
     output = mp_metagga_relax_flow(atoms)
     assert output["static"]["nsites"] == len(atoms)
-    assert output["relax"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax"]["parameters"]["ediffg"] == -0.02
-    assert output["relax"]["parameters"]["encut"] == 680
-    assert output["relax"]["parameters"]["ismear"] == 0
-    assert output["relax"]["parameters"]["sigma"] == 0.05
-    assert output["relax"]["parameters"]["kspacing"] == 0.22
-    assert output["relax"]["parameters"]["pp"] == "PBE"
-    assert output["prerelax"]["parameters"]["gga"] == "PS"
+    assert output["relax"]["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax"]["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax"]["relax2"]["parameters"]["encut"] == 680
+    assert output["relax"]["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax"]["relax2"]["parameters"]["sigma"] == 0.05
+    assert output["relax"]["relax2"]["parameters"]["kspacing"] == 0.22
+    assert output["relax"]["relax2"]["parameters"]["pp"] == "pbe"
+    assert output["prerelax"]["parameters"]["gga"] == "ps"
     assert output["prerelax"]["parameters"]["ismear"] == 0
-    assert output["prerelax"]["parameters"]["pp"] == "PBE"
+    assert output["prerelax"]["parameters"]["pp"] == "pbe"
 
     atoms = bulk("C")
     output = mp_metagga_relax_flow(atoms)
     assert output["static"]["nsites"] == len(atoms)
-    assert output["relax"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax"]["parameters"]["ediffg"] == -0.02
-    assert output["relax"]["parameters"]["encut"] == 680
-    assert output["relax"]["parameters"]["ismear"] == 0
-    assert output["relax"]["parameters"]["kspacing"] == pytest.approx(
+    assert output["relax"]["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax"]["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax"]["relax2"]["parameters"]["encut"] == 680
+    assert output["relax"]["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax"]["relax2"]["parameters"]["kspacing"] == pytest.approx(
         0.28329488761304206
     )
-    assert output["relax"]["parameters"]["pp"] == "PBE"
+    assert output["relax"]["relax2"]["parameters"]["pp"] == "pbe"
     assert output["prerelax"]["parameters"]["ismear"] == 0
-    assert output["prerelax"]["parameters"]["pp"] == "PBE"
+    assert output["prerelax"]["parameters"]["pp"] == "pbe"
 
     atoms = molecule("O2")
     atoms.center(vacuum=10)
@@ -551,13 +554,156 @@ def test_mp_metagga_relax_flow(tmp_path, monkeypatch):
     assert output["static"]["parameters"]["ismear"] == -5
     assert output["static"]["parameters"]["nsw"] == 0
     assert output["static"]["parameters"]["algo"] == "fast"
-    assert output["relax"]["parameters"]["metagga"].lower() == "r2scan"
-    assert output["relax"]["parameters"]["ediffg"] == -0.02
-    assert output["relax"]["parameters"]["encut"] == 680
-    assert output["relax"]["parameters"]["ismear"] == 0
-    assert output["relax"]["parameters"]["kspacing"] == pytest.approx(
+    assert output["relax"]["relax2"]["parameters"]["metagga"].lower() == "r2scan"
+    assert output["relax"]["relax2"]["parameters"]["ediffg"] == -0.02
+    assert output["relax"]["relax2"]["parameters"]["encut"] == 680
+    assert output["relax"]["relax2"]["parameters"]["ismear"] == 0
+    assert output["relax"]["relax2"]["parameters"]["kspacing"] == pytest.approx(
         0.28329488761304206
     )
-    assert output["relax"]["parameters"]["pp"] == "PBE"
+    assert output["relax"]["relax2"]["parameters"]["pp"] == "pbe"
     assert output["prerelax"]["parameters"]["ismear"] == 0
-    assert output["prerelax"]["parameters"]["pp"] == "PBE"
+    assert output["prerelax"]["parameters"]["pp"] == "pbe"
+
+
+def test_mp_gga_relax_job():
+    atoms = bulk("Ni") * (2, 1, 1)
+    atoms[0].symbol = "O"
+    output = mp_gga_relax_job(atoms)
+    assert output["relax2"]["nsites"] == len(atoms)
+    assert output["relax1"]["atoms"].get_chemical_symbols() == ["O", "Ni"]
+    assert output["relax2"]["atoms"].get_chemical_symbols() == ["O", "Ni"]
+    assert output["relax2"]["parameters"] == {
+        "algo": "fast",
+        "ediff": 0.0001,
+        "efermi": "midgap",  # added by copilot
+        "encut": 520,
+        "gamma": True,
+        "ibrion": 2,
+        "isif": 3,
+        "ismear": -5,
+        "ispin": 2,
+        "kpts": [5, 11, 11],
+        "lasph": True,
+        "ldau": True,
+        "ldauj": [0, 0],
+        "ldaul": [0, 2],
+        "ldauprint": 1,  # added by us (sensible)
+        "ldautype": 2,  # added by us (sensible)
+        "ldauu": [0, 6.2],
+        "lmaxmix": 4,
+        "lorbit": 11,
+        "lreal": "auto",
+        "lwave": False,
+        "magmom": [0.6, 0.6],
+        "nelm": 100,
+        "nsw": 99,
+        "prec": "accurate",
+        "sigma": 0.05,
+        "pp": "pbe",
+        "setups": {"O": "", "Ni": "_pv"},
+    }
+
+
+def test_mp_gga_static_job():
+    atoms = bulk("Ni") * (2, 1, 1)
+    atoms[0].symbol = "O"
+    output = mp_gga_static_job(atoms)
+    assert output["nsites"] == len(atoms)
+    assert output["parameters"] == {
+        "algo": "fast",
+        "ediff": 0.0001,
+        "efermi": "midgap",  # added by copilot
+        "encut": 520,
+        "gamma": True,
+        "ismear": -5,
+        "ispin": 2,
+        "kpts": [6, 13, 13],
+        "lasph": True,
+        "lcharg": True,  # modified by us (sensible)
+        "ldau": True,
+        "ldauj": [0, 0],
+        "ldaul": [0, 2],
+        "ldauprint": 1,  # added by us (sensible)
+        "ldautype": 2,  # added by us (sensible)
+        "ldauu": [0, 6.2],
+        "lmaxmix": 4,
+        "lorbit": 11,
+        "lreal": False,
+        "lwave": True,  # modified by us (sensible)
+        "magmom": [0.6, 0.6],
+        "nelm": 100,
+        "nsw": 0,
+        "prec": "accurate",
+        "sigma": 0.05,
+        "pp": "pbe",
+        "setups": {"Ni": "_pv", "O": ""},
+    }
+
+
+def test_mp_gga_relax_flow():
+    atoms = bulk("Ni") * (2, 1, 1)
+    atoms[0].symbol = "O"
+    output = mp_gga_relax_flow(atoms)
+    assert output["static"]["nsites"] == len(atoms)
+    relax_params = {
+        "algo": "fast",
+        "ediff": 0.0001,
+        "efermi": "midgap",  # added by copilot
+        "encut": 520,
+        "gamma": True,
+        "ibrion": 2,
+        "isif": 3,
+        "ismear": -5,
+        "ispin": 2,
+        "kpts": [5, 11, 11],
+        "lasph": True,
+        "ldau": True,
+        "ldauj": [0, 0],
+        "ldaul": [0, 2],
+        "ldauprint": 1,  # added by us (sensible)
+        "ldautype": 2,  # added by us (sensible)
+        "ldauu": [0, 6.2],
+        "lmaxmix": 4,
+        "lorbit": 11,
+        "lreal": "auto",
+        "lwave": False,
+        "magmom": [0.6, 0.6],
+        "nelm": 100,
+        "nsw": 99,
+        "prec": "accurate",
+        "sigma": 0.05,
+        "pp": "pbe",
+        "setups": {"O": "", "Ni": "_pv"},
+    }
+    assert output["relax"]["relax1"]["parameters"] == relax_params
+    assert output["relax"]["relax2"]["parameters"] == relax_params
+    assert output["static"]["parameters"] == {
+        "algo": "fast",
+        "ediff": 0.0001,
+        "efermi": "midgap",  # added by copilot
+        "encut": 520,
+        "gamma": True,
+        "ismear": -5,
+        "ispin": 2,
+        "kpts": [6, 13, 13],
+        "lasph": True,
+        "lcharg": True,  # modified by us (sensible)
+        "ldau": True,
+        "ldauj": [0, 0],
+        "ldaul": [0, 2],
+        "ldauprint": 1,  # added by us (sensible)
+        "ldautype": 2,  # added by us (sensible)
+        "ldauu": [0, 6.2],
+        "lmaxmix": 4,
+        "lorbit": 11,
+        "lreal": False,
+        "lwave": True,  # modified by us (sensible)
+        "magmom": [0.6, 0.6],
+        "nelm": 100,
+        "nsw": 0,
+        "prec": "accurate",
+        "sigma": 0.05,
+        "pp": "pbe",
+        "setups": {"Ni": "_pv", "O": ""},
+    }